Extract RetryError, fix desktop-app and TanStack Router fallout
- Add @everr/ui/components/retry-error and adopt it in traces detail/list
  pages and the logs explorer panel (logs gains a Retry button to match)
- Rename routes/api/cli/test-utils.ts to -test-utils.ts so TanStack Router
  ignores it (was warning 'does not export a Route')
- Move DEFAULT_TIME_RANGE and TimeRange into @everr/ui/lib/time-range so
  consumers no longer pull them through the JSX-bearing time-range-picker
  re-export (was breaking the desktop app at runtime)
- Point desktop-app traces-page at @everr/ui directly for the time-range
  and refresh-picker helpers that telemetry-explorer no longer re-exports
