+++ packages/telemetry-explorer/src/traces/ui/trace-results-list.tsx
@@ -5,6 +5,7 @@ import {
   EmptyHeader,
   EmptyTitle,
 } from "@everr/ui/components/empty";
+import { RetryError } from "@everr/ui/components/retry-error";
 import { Skeleton } from "@everr/ui/components/skeleton";
 import { formatDuration } from "@everr/ui/lib/formatting";
 import type { UseQueryResult } from "@tanstack/react-query";
@@ -55,7 +56,8 @@ export function TraceResultsList({
   if (query.isPending) return showSkeleton ? <ResultsSkeleton /> : null;
   if (query.isError) {
     return (
-      <ErrorState
+      <RetryError
+        title="Failed to load traces"
         message={(query.error as Error).message}
         onRetry={() => query.refetch()}
       />
@@ -157,26 +159,6 @@ function ResultsSkeleton() {
         <Skeleton key={i} className="h-10 w-full" />
       ))}
     </div>
-  );
-}
-
-function ErrorState({
-  message,
-  onRetry,
-}: {
-  message: string;
-  onRetry: () => void;
-}) {
-  return (
-    <Empty>
-      <EmptyHeader>
-        <EmptyTitle>Failed to load traces</EmptyTitle>
-        <EmptyDescription>{message}</EmptyDescription>
-      </EmptyHeader>
-      <Button variant="outline" size="sm" onClick={onRetry}>
-        Retry
-      </Button>
-    </Empty>
   );
 }
